Move bash utils into a lib.
This actually matters quite a lot since these get used in scripts that
sync forward and backward in git. If they get sourced at the start, then
the scripts themselves will all work the way they do in the local client
(or latest check-in) when the script is run. Previously, they would
reflect the baseline bash source code when running the baseline stats.